Issue #17429: Relocate inputs to compilable path for IT Module 2
