+++ config/jsoref-spellchecker/whitelist.words
@@ -1107,7 +1107,6 @@ Popup
 Postgresql
 powershell
 prebuilts
-prefs
 PRESIZE
 prettyprint
 printf
