--- src/it/resources-noncompilable/org/checkstyle/suppressionxpathfilter/coding/patternvariableassignment/InputXpathPatternVariableAssignmentClass.java
@@ -1,16 +1,0 @@
-// non-compiled with javac: Compilable with Java16
-package org.checkstyle.suppressionxpathfilter.coding.patternvariableassignment;
-
-public class InputXpathPatternVariableAssignmentClass {
-  public void foo() {
-    AnonymousClass annClass = new AnonymousClass() {
-      public void test(Object obj) {
-        if (obj instanceof Integer x) {
-          x = 3; // warn
-        }
-      }
-    };
-
-  }
-
-}
