--- src/it/resources-noncompilable/org/checkstyle/suppressionxpathfilter/naming/staticvariablename/InputXpathStaticVariableNameNoAccessModifier.java
@@ -1,8 +1,0 @@
-// non-compiled with javac: compilable with java21
-package org.checkstyle.suppressionxpathfilter.staticvariablename;
-
-public class InputXpathStaticVariableNameNoAccessModifier {
-    {
-       static int NUM3; //warn
-    }
- }
