+++ src/it/resources/org/checkstyle/suppressionxpathfilter/coding/whenshouldbeused/InputXpathWhenShouldBeUsedNested.java
@@ -1,4 +1,4 @@
-// non-compiled with javac: Compilable with Java21
+// Java21
 package org.checkstyle.suppressionxpathfilter.coding.whenshouldbeused;
 
 public class InputXpathWhenShouldBeUsedNested {
