+++ src/it/resources/org/checkstyle/suppressionxpathfilter/coding/whenshouldbeused/InputXpathWhenShouldBeUsedSimple.java
@@ -1,4 +1,4 @@
-// non-compiled with javac: Compilable with Java21
+// Java21
 package org.checkstyle.suppressionxpathfilter.coding.whenshouldbeused;
 
 public class InputXpathWhenShouldBeUsedSimple {
