+++ src/it/resources/org/checkstyle/suppressionxpathfilter/naming/patternvariablename/InputXpathPatternVariableNameThree.java
@@ -1,0 +1,9 @@
+// java21
+package org.checkstyle.suppressionxpathfilter.naming.patternvariablename;
+
+public class InputXpathPatternVariableNameThree {
+    void MyClass(Object o1) {
+        if (o1 instanceof String STR) { // warning
+        }
+    }
+}
