+++ src/it/resources/org/checkstyle/suppressionxpathfilter/naming/staticvariablename/InputXpathStaticVariableNameNoAccessModifier.java
@@ -1,0 +1,8 @@
+// java21
+package org.checkstyle.suppressionxpathfilter.naming.staticvariablename;
+
+public class InputXpathStaticVariableNameNoAccessModifier {
+    static int NUM3; //warn
+    {
+    }
+}
